<refactor> refoctoring dags and coin directory
- 기존의 coin_flow.py 를 시총이 가장 큰 btc 그리고 eth , ada 정도로
추가하였습니다.

- 확장성을 조금 더 고려하여 coin 디렉토리에 있는 함수들에 ticker 인자를
받을 수 있도록 수정하였습니다.

- git commit 할 때 black이 아닌 flake8때문에 commit이 안됐던거여서
pre-commit-config.yaml 파일에 dags 는 flake8 에서 제외되도록 수정하였습니다.

- close #98
