+++ .pre-commit-config.yaml
@@ -1,6 +1,6 @@
 repos:
 -   repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v4.2.0
+    rev: v4.3.0
     hooks:
     -   id: trailing-whitespace
     -   id: check-yaml
@@ -14,3 +14,4 @@ repos:
     rev: 3.9.2
     hooks:
     - id: flake8
+      exclude: ^dags
